+++ src/domain/index.ts
@@ -1,2 +1,3 @@
 export * from './Post/postTypes';
 export * from './Post/postService';
+export * from './Post/useCases/usePostList';
